+++ src/SupportTools/Public/Invoke-IncidentResponse.ps1
@@ -16,8 +16,14 @@ function Invoke-IncidentResponse {
         [switch]$Simulate,
         [Parameter(Mandatory = $false)]
         [switch]$Explain
+        ,[Parameter(Mandatory = $false)]
+        [object]$Logger
+        ,[Parameter(Mandatory = $false)]
+        [object]$TelemetryClient
+        ,[Parameter(Mandatory = $false)]
+        [object]$Config
     )
     process {
-        Invoke-ScriptFile -Name 'Invoke-IncidentResponse.ps1' -Args $Arguments -TranscriptPath $TranscriptPath -Simulate:$Simulate -Explain:$Explain
+        Invoke-ScriptFile -Logger $Logger -TelemetryClient $TelemetryClient -Config $Config -Name 'Invoke-IncidentResponse.ps1' -Args $Arguments -TranscriptPath $TranscriptPath -Simulate:$Simulate -Explain:$Explain
     }
 }